[#18] Removed unused gems from Gemfile
The following gems were removed either because they are not used anymore
or because they are not an app dependency.

For example, we still use foreman but this should not be installed
through the gemfile. Run `gem install foreman` to install it.

rails-erd
ruby-graphviz
ruby-prof
bluecloth
choice
foreman
gon
annotate